replace old sheet with gesture sheet. api remains the same
